src/requirements.txt
matplotlib
numpy
scikit-learn
scikit-image
pillow
scipy
tqdm
pybind11
torchvision